Displaying the number of times an author appears first/last on a paper. Tests for the new feature. Some rectification for the things caused problems on my laptop (if that causes other problems for you, I will change them to default version)

+++ src/comp62521/statistics/average.py
@@ -11,8 +11,8 @@ def median(X):
         return 0
     L = sorted(X)
     if n % 2:
-        return L[n / 2]
-    return mean(L[(n / 2) - 1:(n / 2) + 1])
+        return L[int(n / 2)]
+    return mean(L[int((n / 2) - 1):int((n / 2) + 1)])
 
 
 def mode(X):
